Removing references to semantic model
Semantic model not fully supported for now. We need to add the logic for
using and not using because at the moment, if no semantic model is
provided, one is generated and is not good.

Once ready, we need to have components be ready to accept a semantic
model and use it in case. Fall back strategies are to be applied in case
semantic model is not available.