Removing Source
`Source` was initially introduced as a singleton component for handling
common creating of the semantic model.

In previous changes, the code in helpers relying on the semantic model
has been guarded with conditionals in order to check whether a semantic
model is available, and in the opposite case, trigger a guess logic for
handling the task.

So no more need for `Source`. This also removes a lot of calls that were
being made to `BuildSemanticModel` which were causing an approximate
application slowdown by 15% by profiling.

This closes #2