Semantic model full extent - Using full names in references
This pack wires up the semantic model to a full extent. Now we wire up
semantics in all places and are able to emit full names in signatures.

Tests are not passing a semantic model, thus many of them will emit just
names, no full names.

At the moment the ScriptSharp definition generator is emitting full
names. The normal translator is also capable of emitting full names,
though the runner has not really been tested in a E2E scenario.

Problem: This pack breaks the transformer wihch had to be deactivated.
The transformer is rearranging the AST, problem is that we are not
updating the semantic model, thus references to nodes are not found in
the semantic model which complains about node not found within the
semantic model. Future packs will fix this.

Targets #19