Semantic alignment - Foundamental achieving
Achieving semantic alignment in order to use the Semantic Model the most
possible after modifying the AST.

This pack introduced 2 points for achieveing semantic alignment:
- Semantic update after AST transformation.
- Semantic alignment at helper level. This feature makes use of the
Semantic Model in helpers when having references to types, they are
still in the old namespace. Detecting in semantic model whether the type
has a `ScriptNamespace` and acting accordingly.

Targets #22