Ability to render references for definition generation
This pack targets the ability to emit specific references in the emitted
code when generating definition files.

A new component has been introduced in this pack and has changed the
flow of data in the `Executable` namespace and derived:
`ConversionArguments`. The problem is that `ConversionProvider` might
take all possible inputs, instead of having them in the singature of the
delegate, we make the delegate accept a new type which has all possible
inputs.

The feature is impacting the `ProgramWrapper` specific ti ScriptSharp
definition generation projects only.

### Notes
Assembly and References are still not wired-up to wrappers and
executables (for example build task for ScriptSharo definition
generation). Next packs will have to target this.

Closes #21