Introducing project level separation betweern ScriptSharp and ScriptSharp.Definition
The product has been separated in order to define the logic for
converting CS into TS from the logic to generate D.TS. Problem is that
in future we might want to introduce specifc translation specific to CS
code written by targeting ScriptSharp. So projects have been separated
according to this new division.

Closes #32