Wriring up Reflection to execution layer
Tests added by extending the build task example project.
Extended runner in ScriptSharp for keeping into account the assembly.

Attention: The runner does not work because the assembly is not loaded
into the inspection only context and all reference errors are surfacing.
This must be fixed in later pack.

Targets #41