Adding basic support for object creation expression
Introduding support for object creation expressions with renderers and
tests targeting the most basic scenario: parameterless object
construction.

This is related to #41 as the generation of the AST in `ASTBuilder`
needs the creation of empty block codes where we emit `throw new
NotImplementedException()`. SO far the object creation expression was
not supported, thus leading to issues here.