Removing raw assembly requirement for reflection
The generation of AST from assembly was failing because the semantic
model was computed on the original assembly. Because of the
ScriptNamespace, the AST generator will change the AST from the original
one in the assembly, this was causing a misalignment in the semantic
model being used and, at AST walking time, the ScriptNamespace attribute
was detected wrongly.

Fixed the issue and also got rid of the raw assembly in reflcetion.

Targets #41