Naming: Visibility vs. Modifiers
In a lot of places in the codebase, the handling of modififers for type
definitions and member definitions started just with visibility. Later
on this got extended to support modifiers in general, however the code
was still referencing modifiers as `visibility`.

Changed names in order to reflect the proper semantics.