Fixing missing static modifier in definitions
Static modifiers was not rendered in ScriptSharp definitions because
translation units, attempting to ignore the `public` modifier, were
actually ignoring all modifiers. This is an issue that arises from our
intention to extend the `visibility` to `modifiers`. In some places we
still consider modifiers to only be `public`, `protected` or `internal`
while `static` and other modifiers may apply.

FIxes #46