Introducing StringRendering Test Framework
Initial code (working) for comparing generated texts. This change does
not add new tests to be executed, it just defines an example test
(failing) in order to build up on that in future.

Next step is to add `.Tests` projects in renderers and have more
generated code to compare and add more tests.

Targets #47